various PEP8 and dtype fixes

+++ madmom/features/beats.py
@@ -286,7 +286,7 @@ def detect_beats(activations, interval, look_aside=0.2):
     positions = []
     recursive(pos)
     # return indices (as floats, since they get converted to seconds later on)
-    return np.array(positions).astype(np.float)
+    return np.array(positions, dtype=np.float32)
 
 
 # default values for beat tracking
